use memory allocator in byte array dictionary
